+++ ContentBlocker/BlockerListLoader.swift
@@ -10,30 +10,28 @@ import Foundation
 import MobileCoreServices
 
 protocol BlockerListLoader {
-    func loadBlockerList()
-    func loadBlockerListSuccess(withReturningItems returningItems: [NSExtensionItem])
-    func loadBlockerListFailed(withError error: Error)
+    func loadBlockerList(completion: @escaping (([NSExtensionItem]?, Error?) -> Void))
 }
 
 extension BlockerListLoader {
-    func loadBlockerList() {
+    func loadBlockerList(completion: @escaping (([NSExtensionItem]?, Error?) -> Void)) {
         let session = URLSession(configuration: .default)
         guard let url = URL(string: "https://raw.githubusercontent.com/ethanhuang13/blahker/master/Blahker.safariextension/blockerList.json") else {
-            loadBlockerListFailed(withError: ContentBlockerRequestHandlerError.createBlockerListUrlFailed)
+            completion(nil, ContentBlockerRequestHandlerError.createBlockerListUrlFailed)
             return
         }
 
         let request = URLRequest(url: url)
         let task = session.dataTask(with: request) { (data, response, connectionError) -> Void in
             guard let data = data as? NSSecureCoding else {
-                self.loadBlockerListFailed(withError: ContentBlockerRequestHandlerError.loadBlockerListFileFailed)
+                completion(nil, ContentBlockerRequestHandlerError.loadBlockerListFileFailed)
                 return
             }
 
             let attachment = NSItemProvider(item: data, typeIdentifier: kUTTypeJSON as String)
             let item = NSExtensionItem()
             item.attachments = [attachment]
-            self.loadBlockerListSuccess(withReturningItems: [item])
+            completion([item], nil)
         }
 
         task.resume()